Extract of subset of Vigibase

Create a subset of the VigiBase ECL datasets

Usage

tb_subset(
  wd_in,
  wd_out,
  subset_var = c("drecno", "medprod_id", "meddra_id", "age"),
  sv_selection,
  rm_suspdup = TRUE
)

Arguments

wd_in

Source directory pathway (character)

wd_out

Output directory pathway (character)

subset_var

One of "drecno", "medprod_id", "meddra_id", "age"

sv_selection

A named list or a vector containing the appropriate ids (according to the method, see details)

rm_suspdup

A logical. Should suspected duplicates be removed? TRUE by default

Value

Parquet files in the output directory. All files from a vigibase ECL main folder are returned subsetted (including suspectedduplicates).

Details

You must select a subset variable with subset_var and provide an appropriate list according to this variable in sv_selection. Available subset_var :

drecno will use Drug Record Number (DrecNo), from WHO Drug, and will subset from drug (see get_drecno()).

medprod_id will use MedicinalProd_Id, also from drug. May be useful if requesting from ATC classes. (see get_atc_code()).

meddra_id will use MedDRA_Id, subset from adr. (see get_llt_soc() or See get_llt_smq()).

age will use AgeGroup from demo. See below.

Age groups ids are as follows:

1 0 - 27 days

2 28 days to 23 months

3 2 - 11 years

4 12 - 17 years

5 18 - 44 years

6 45 - 64 years

7 65 - 74 years

8 >= 75 years

9 Unknown

Example: To work with patients aged 18 to 74, provide c(5, 6, 7) as sv_selection.

Use dt_parquet() to load the tables afterward.

See also

Examples

## Extract all colitis cases

# Locate your input directory
# wd_in <- "../vigibase_ecl/main/"
# Choose a (close) directory to export to (it doesnt have to exist before you call it)
# wd_out <- "../vigibase_ecl/main/subset/"

# For the example, we create tables in a directory that should be replaced by
# your own directory containing the entire database.

wd_in <- tempdir()

# --- technical steps ---- #

library(data.table)

mini_data <- rlang::list2(
  demo =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      AgeGroup = c(1, 2, 7, 9)
    ),
  drug =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      Drug_Id = c("d1", "d2", "d3", "d4"),
      DrecNo = c(133138448, 133138448, 111841511, 111841511),
      MedicinalProd_Id = c(25027716, 97354576, 104264760, 37484408)
    ),
  adr  =
    data.table(
      UMCReportId = c(1, 2, 3, 4),
      Adr_Id = c("a1", "a2", "a3", "a4"),
      MedDRA_Id = c(110049083, 31672047, 146319904, 72535511)
    ),
  link =
    data.table(
      Drug_Id = c("d1", "d2", "d3", "d4"),
      Adr_Id = c("a1", "a2", "a3", "a4")
    ),
  srce =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  ind  =
    data.table(
      Drug_Id = c("d1", "d2", "d3", "d4")
    ),
  out  =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  followup =
    data.table(
      UMCReportId = c(1, 2, 3, 4)
    ),
  suspdup =
    data.table(
      UMCReportId = c(3),
      SuspectedduplicateReportId = c(4)
    )
)

  purrr::iwalk(
    mini_data,
    function(dataset, name)
      arrow::write_parquet(
        dataset |>
          arrow::as_arrow_table(),
        sink = paste0(wd_in, "/", name, ".parquet")
      )

  )

# back to tb_subset, you should select a subset_var and corresponding data

# Subset on adr colitis codes

adr_llt <-
 list(
   colitis = "Colitis"
   ) |>
   get_llt_soc(term_level = "pt", meddra_)
#> 
#> ── get_llt_soc() ───────────────────────────────────────────────────────────────
#> 
#> ── ✔ Matched reactions at `pt` level (number of codes) ──
#> 
#> → `colitis`: "Colitis (25)"
#> 
#> ℹ Set `verbose` to FALSE to suppress this section.
#> 

wd_out <- paste0(wd_in, "/", "colitis_subset", "/")

tb_subset(wd_in, wd_out,
          subset_var = "meddra_id",
          sv_selection = adr_llt)
#> ── tb_subset() ─────────────────────────────────────────────────────────────────
#> ℹ Subsetting VigiBase tables.
#> ■■■■■                             15% | 2ms | Reading source tables 
#> ■■■■■■■■■■                        30% | 16ms | Picking subset 
#> ■■■■■■■■■■■■■■■                   45% | 69ms | Collect case list 
#> ■■■■■■■■■■■■■■■■                  50% | 76ms | Collect drug list 
#> ■■■■■■■■■■■■■■■■■■■               60% | 97ms | Apply case level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | 196ms | Apply drug level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | 253ms | Done 
#> ✔ Subset successful
#> 
#> ℹ   Table  | Number of rows in subset
#> →   demo   | 1
#> →   adr    | 1
#> →   out    | 1
#> →   srce   | 1
#> → followup | 1
#> →   drug   | 1
#> →   link   | 1
#> →   ind    | 1
#> 

# Subset on drug codes

 d_drecno <-
   list(
    ipi = "ipilimumab") |>
    get_drecno(mp = mp_)
#> ── get_drecno() ────────────────────────────────────────────────────────────────
#> 
#> ── `d_sel`: Matching drugs ──
#> 
#> ── ✔ Matched drugs 
#> 
#> → `ipi`: "ipilimumab" and "ipilimumab;nivolumab"
#> 
#> ℹ Set `verbose` to FALSE to suppress this section.
#> 
#> ────────────────────────────────────────────────────────────────────────────────

wd_out <- paste0(wd_in, "/", "nivolumab_subset", "/")

tb_subset(wd_in, wd_out,
          subset_var = "drecno",
          sv_selection = d_drecno)
#> 
#> ── tb_subset() ─────────────────────────────────────────────────────────────────
#> ℹ Subsetting VigiBase tables.
#> ■■■■■                             15% | 1ms | Reading source tables 
#> ■■■■■■■■■■                        30% | 14ms | Picking subset 
#> ■■■■■■■■■■■■■■■                   45% | 47ms | Collect case list 
#> ■■■■■■■■■■■■■■■■                  50% | 54ms | Collect drug list 
#> ■■■■■■■■■■■■■■■■■■■               60% | 75ms | Apply case level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | 170ms | Apply drug level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | 229ms | Done 
#> ✔ Subset successful
#> 
#> ℹ   Table  | Number of rows in subset
#> →   demo   | 2
#> →   adr    | 2
#> →   out    | 2
#> →   srce   | 2
#> → followup | 2
#> →   drug   | 2
#> →   link   | 2
#> →   ind    | 2
#> 

 # Subset on age > 65 year-old

 sv_selection <-
    c(7, 8)

wd_out <- paste0(wd_in, "/", "more_than_65_subset", "/")

tb_subset(wd_in, wd_out,
          subset_var = "age",
          sv_selection = sv_selection)
#> ── tb_subset() ─────────────────────────────────────────────────────────────────
#> ℹ Subsetting VigiBase tables.
#> ■■■■■                             15% | 1ms | Reading source tables 
#> ■■■■■■■■■■                        30% | 14ms | Picking subset 
#> ■■■■■■■■■■■■■■■                   45% | 44ms | Collect case list 
#> ■■■■■■■■■■■■■■■■                  50% | 51ms | Collect drug list 
#> ■■■■■■■■■■■■■■■■■■■               60% | 71ms | Apply case level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | 167ms | Apply drug level subset 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | 224ms | Done 
#> ✔ Subset successful
#> 
#> ℹ   Table  | Number of rows in subset
#> →   demo   | 1
#> →   adr    | 1
#> →   out    | 1
#> →   srce   | 1
#> → followup | 1
#> →   drug   | 1
#> →   link   | 1
#> →   ind    | 1
#>