Create main VigiBase ECL tables

Transform VigiBase .txt files to .parquet files.

Usage

tb_vigibase(path_base, path_sub)

Arguments

path_base

Character string, a directory containing vigibase txt tables. It is also the output directory.

path_sub

Character string, a directory containing subsidiary tables.

Value

.parquet files of all main tables into the path_base directory: demo, adr, drug, link, ind, out, srce, followup, and the suspdup (suspected duplicates) table. Check ?demo_ for more information on the tables.

The link table is augmented with tto_mean and range, to analyze time to onset according to WHo's recommendations (see vignette("descriptive").

.parquet files of all other subsidiary tables into the path_sub directory: AgeGroup, Dechallenge, Dechallenge2, Frequency, Gender, Notifier, Outcome, Rechallenge, Rechallenge2, Region, RepBasis, ReportType, RouteOfAdm, Seriousness, and SizeUnit.

.parquet files into the path_base directory (including suspected duplicates tables). Some columns are returned as integer (UMCReportId, Drug_Id, MedicinalProd_Id, Adr_Id, MedDRA_Id), and some columns as numeric (TimeToOnsetMin, TimeToOnsetMax) All other columns are character.

Details

Vigibase Extract Case Level is delivered as zipped text files, that you should transform to a more efficient format. Parquet format from arrow has many advantages: It works with out-of-memory data, which makes it possible to process Vigibase tables on a computer with not-so-much RAM. It is also lightweighted and standard across different langages. The function also creates variables in each table. The suspectedduplicates table will be added to the base directory. Use dt_parquet() to load the tables afterward.

See also

Examples

# --- Setting to try out ---- ####

f_sets <-
 list(
   DEMO.txt = data.frame(f0= c("96548661   32194501051119460820")),
   DRUG.txt = data.frame(f0 =
             c("70548965   8          4901354    064392080055011    31- 806")
               ),
   LINK.txt = data.frame(
   f0 = c("2          654654     51---0.78991   0.98745    ",
          "2          456456     51---6.98789   -          ")),
   FOLLOWUP.txt = data.frame(f0 = c("0548978    0254687    ",
                                    "7568798    4565321    ")),
   ADR.txt = data.frame(f0 = c("96570161   14         100474561",
                               "70578465   17         145078144")),
   OUT.txt = data.frame(f0 = c("70547815   - N",
                               "96575661   - Y")),
   SRCE.txt = data.frame(f0 = c("4898765    1 ",
                                "9804562    1 ")),
   IND.txt = data.frame(# 266 length
     f0 = paste0("780954     Cutaneous diseases due to other mycobacteria",
                 rep(" ", 211))
     ),
   SUSPECTEDDUPLICATES.txt = data.frame(f0 = c("789054     789542     ",
                                               "780546     654352     ")),
   AgeGroup_Lx.txt = data.frame(f0 = c("1An age range             ")),
   Dechallenge_Lx.txt = data.frame(f0 = paste0("1Some drug action",
                                               rep(" ", 237))),
   Dechallenge2_Lx.txt = data.frame(f0 = paste0("1Some outcome occurring",
                                                rep(" ", 231))),
   Frequency_Lx.txt =
     data.frame(f0 =
                  paste0("123Some frequency of administration",
                         rep(" ", 221))),
   Gender_Lx.txt = data.frame(f0 = paste0("1Some gender",
                                          rep(" ", 242))),
   Notifier_Lx.txt = data.frame(f0 = paste0("1 Some notifier",
                                            rep(" ", 240))),
   Outcome_Lx.txt = data.frame(f0 = paste0("1Some outcome",
                                           rep(" ", 241))),
   Rechallenge_Lx.txt = data.frame(f0 = paste0("1A rechallenge action",
                                               rep(" ", 60))),
   Rechallenge2_Lx.txt =
     data.frame(f0 = paste0("1A reaction recurrence status",
                            rep(" ", 36))),
   Region_Lx.txt = data.frame(f0 = paste0("1A world region",
                                          rep(" ", 36))),
   RepBasis_Lx.txt = data.frame(f0 = paste0("1A reputation basis",
                                            rep(" ", 32))),
   ReportType_Lx.txt = data.frame(f0 = paste0("1A type of report",
                                              rep(" ", 237))),
   RouteOfAdm_Lx.txt = data.frame(f0 = paste0("1 A route of admnistration",
                                              rep(" ", 56))),
   Seriousness_Lx.txt = data.frame(f0 = paste0("1 Some seriousness criteria",
                                               rep(" ", 224))),
   SizeUnit_Lx.txt = data.frame(f0 = paste0("1 A dosing unit",
                                            rep(" ", 66)))
 )

tmp_folder <- tempdir()

path_base <- paste0(tmp_folder, "/", "main", "/")

if(!dir.exists(path_base)) {dir.create(path_base)}

path_sub  <- paste0(tmp_folder, "/", "sub",  "/")

if(!dir.exists(path_sub)) {dir.create(path_sub)}

purrr::iwalk(f_sets, function(d_, name_){
  if(name_ %in%
     c("SUSPECTEDDUPLICATES.txt",
       "AgeGroup_Lx.txt",
       "Dechallenge_Lx.txt",
       "Dechallenge2_Lx.txt",
       "Frequency_Lx.txt",
       "Gender_Lx.txt",
       "Notifier_Lx.txt",
       "Outcome_Lx.txt",
       "Rechallenge_Lx.txt",
       "Rechallenge2_Lx.txt",
       "Region_Lx.txt",
       "RepBasis_Lx.txt",
       "ReportType_Lx.txt",
       "RouteOfAdm_Lx.txt",
       "Seriousness_Lx.txt",
       "SizeUnit_Lx.txt")
  ){
    write.table(d_, file = paste0(path_sub, name_),
      row.names = FALSE, quote = FALSE, col.names = FALSE)
  } else {
    write.table(d_, file = paste0(path_base, name_),
      row.names = FALSE, quote = FALSE, col.names = FALSE)
  }
})

 # ---- Running tb_vigibase

 tb_vigibase(path_base = path_base,
        path_sub  = path_sub)
#> 
#> ── tb_vigibase() ───────────────────────────────────────────────────────────────
#> ℹ Creating vigibase tables.
#> This process must only be done once per database version.
#> It can take up to 30minutes.
#> ■■                                 3% | 1ms | Read DEMO.txt 
#> ■■■                                6% | 13ms | Split demo 
#> ■■■■■                             12% | 55ms | Write demo.parquet 
#> ■■■■■■                            16% | 208ms | Read DRUG.txt 
#> ■■■■■■■                           20% | 218ms | Split drug 
#> ■■■■■■■■■                         27% | 281ms | Write drug.parquet 
#> ■■■■■■■■■■                        30% | 422ms | Read FOLLOWUP.txt 
#> ■■■■■■■■■■■                       32% | 433ms | Split followup 
#> ■■■■■■■■■■■                       34% | 468ms | Write followup.parquet 
#> ■■■■■■■■■■■■                      36% | 601ms | Read ADR.txt 
#> ■■■■■■■■■■■■■                     39% | 612ms | Split adr 
#> ■■■■■■■■■■■■■■                    42% | 653ms | Write adr.parquet 
#> ■■■■■■■■■■■■■■                    43% | 661ms | Read OUT.txt 
#> ■■■■■■■■■■■■■■■                   45% | 670ms | Split out 
#> ■■■■■■■■■■■■■■■                   46% | 706ms | Write out.parquet 
#> ■■■■■■■■■■■■■■■                   47% | 844ms | Read SRCE.txt 
#> ■■■■■■■■■■■■■■■                   48% | 856ms | Split srce 
#> ■■■■■■■■■■■■■■■■                  49% | 889ms | Write srce.parquet 
#> ■■■■■■■■■■■■■■■■                  50% | 1s | Read LINK.txt 
#> ■■■■■■■■■■■■■■■■■                 52% | 1s | Split link (longest step) 
#> ■■■■■■■■■■■■■■■■■■■■■             68% | 1.2s | Write link.parquet 
#> ■■■■■■■■■■■■■■■■■■■■■■            70% | 1.3s | Read IND.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■           72% | 1.4s | Split ind 
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | 1.4s | Write ind.parquet 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■         80% | 1.5s | Read SUSPECTEDDUPLICATES.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        82% | 1.5s | Split suspdup 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        84% | 1.6s | Write suspdup.parquet 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       85% | 1.7s | Process AgeGroup_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       86% | 1.7s | Process Dechallenge_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | 1.8s | Process Dechallenge2_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       88% | 1.8s | Process Frequency_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      89% | 1.8s | Process Gender_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      90% | 1.9s | Process Notifier_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91% | 1.9s | Process Outcome_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     92% | 1.9s | Process Rechallenge_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     93% | 1.9s | Process Rechallenge2_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     94% | 2s | Process Region_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■     95% | 2s | Process RepBasis_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% | 2s | Process ReportType_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    97% | 2.1s | Process RouteOfAdm_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    98% | 2.1s | Process Seriousness_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■   99% | 2.1s | Process SizeUnit_Lx.txt 
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | 2.1s | Done